Add files via upload
In this project I analysed a data base with medical prescriptions with two or more psycothropic phamarceuticals. I analysed it by prevalence, demographics and also differenciated two kinds of prescriptions, contentious or not.  A prescription is considered contentious if 3 or more pharmaceuticals in the prescription are of the same class, or 2 or more are of the same subclass.